chore: remove zustand socket import and reorganize socket imports
