+++ src/api/index.ts
@@ -6,5 +6,3 @@ export * from './types';
 export * from './QueryClientProvider';
 export * from './fetcher';
 export * from './refreshAccessToken';
-
-export * from './socket/useSocket';
