+++ src/pages/GameSelectPage/GameMatchingPage/CustomMatchingPage/CustomTournament/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useRef, useState } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
-import { useSocket, useUsersMe, CustomCreateResponse, WaitingRoomUpdateResponse } from '@/api';
-import { useGameInviteSocket } from '@/api/socket';
+import { useUsersMe, CustomCreateResponse, WaitingRoomUpdateResponse } from '@/api';
+import { useGameInviteSocket, useSocket } from '@/api/socket';
 import { Flex } from '@/components/system';
 import { BackButton } from '@/components/ui';
